+++ StripePayments/StripePayments/Source/API Bindings/Models/PaymentMethods/STPPaymentMethodEnums.swift
@@ -118,6 +118,8 @@ import Foundation
     case goPay
     /// A MoMo payment method
     case momo
+    /// A GCash payment method
+    case gcash
     /// An unknown type.
     case unknown
 
@@ -232,6 +234,8 @@ import Foundation
             return "Scalapay"
         case .momo:
             return "MoMo"
+        case .gcash:
+            return "GCash"
         case .goPay:
             return "GoPay"
         case .cardPresent,
@@ -351,6 +355,8 @@ import Foundation
             return "scalapay"
         case .momo:
             return "momo"
+        case .gcash:
+            return "gcash"
         case .goPay:
             return "gopay"
         }