+++ StripePayments/StripePayments/Source/API Bindings/Models/PaymentMethods/STPPaymentMethodParams.swift
@@ -129,6 +129,8 @@ public class STPPaymentMethodParams: NSObject, STPFormEncodable {
     @objc public var scalapay: STPPaymentMethodScalapayParams?
     /// If this is a GoPay PaymentMethod, this contains additional details.
     @objc public var goPay: STPPaymentMethodGoPayParams?
+    /// If this is a GCash PaymentMethod, this contains additional details.
+    @objc public var gcash: STPPaymentMethodGCashParams?
     /// If this is a MoMo PaymentMethod, this contains additional details.
     @objc public var momo: STPPaymentMethodMomoParams?
 
@@ -928,6 +930,24 @@ public class STPPaymentMethodParams: NSObject, STPFormEncodable {
         self.metadata = metadata
     }
 
+    /// Creates params for a GCash PaymentMethod.
+    /// - Parameters:
+    ///   - gcash:       An object containing additional GCash details.
+    ///   - billingDetails: Billing information associated with the PaymentMethod.
+    ///   - metadata:       Additional information to attach to the PaymentMethod.
+    @objc
+    public convenience init(
+        gcash: STPPaymentMethodGCashParams,
+        billingDetails: STPPaymentMethodBillingDetails?,
+        metadata: [String: String]?
+    ) {
+        self.init()
+        self.type = .gcash
+        self.gcash = gcash
+        self.billingDetails = billingDetails
+        self.metadata = metadata
+    }
+
     /// Creates params for a MoMo PaymentMethod.
     /// - Parameters:
     ///   - momo:       An object containing additional MoMo details.
@@ -999,6 +1019,7 @@ public class STPPaymentMethodParams: NSObject, STPFormEncodable {
             NSStringFromSelector(#selector(getter: sequra)): "sequra",
             NSStringFromSelector(#selector(getter: scalapay)): "scalapay",
             NSStringFromSelector(#selector(getter: goPay)): "gopay",
+            NSStringFromSelector(#selector(getter: gcash)): "gcash",
             NSStringFromSelector(#selector(getter: momo)): "momo",
             NSStringFromSelector(#selector(getter: link)): "link",
             NSStringFromSelector(#selector(getter: radarOptions)): "radar_options",
@@ -1463,6 +1484,8 @@ extension STPPaymentMethodParams {
             scalapay = STPPaymentMethodScalapayParams()
         case .goPay:
             goPay = STPPaymentMethodGoPayParams()
+        case .gcash:
+            gcash = STPPaymentMethodGCashParams()
         case .momo:
             momo = STPPaymentMethodMomoParams()
         case .cardPresent, .paynow, .zip, .konbini, .promptPay, .mbWay, .bizum: